requirements.txt
asgiref==3.7.2
Django
django-jazzmin @ git+https://github.com/farridav/django-jazzmin.git@586e81f204033add57b67f657e43681801356a9a
django-phonenumber-field==7.3.0
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.2
phonenumbers==8.13.27
pillow==10.2.0
# psycopg2==2.9.9
psycopg2-binary==2.9.9
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
setuptools==69.0.3ss
sqlparse==0.4.4
tzdata==2023.4
uritemplate==4.1.1
django-cors-headers